+++ src/App.tsx
@@ -4,18 +4,20 @@ import { AppHeader } from './components/header/AppHeader.tsx';
 import { AppLayout } from './components/layout/AppLayout.tsx';
 
 import './App.css';
-import { CardFront } from './components/cards/CardFront.tsx';
-import { CardBack } from './components/cards/CardBack.tsx';
+import { NewCardFront } from './components/cards/NewCardFront.tsx';
+import { EditCardBack } from './components/cards/EditCardBack.tsx';
 import { SingleCard } from './components/cards/SingleCard.tsx';
 import { CardsList } from './components/cards/CardsList.tsx';
+import { EditCardFront } from './components/cards/EditCardFront.tsx';
 
 function App() {
   return (
     <AppLayout>
       <AppHeader />
-      <CardsList />
-      {/*<CardFront />*/}
-      {/*<CardBack />*/}
+      {/*<EditCardFront />*/}
+      {/*<CardsList />*/}
+      <NewCardFront />
+      {/*<EditCardBack />*/}
     </AppLayout>
   );
 }
