+++ src/components/cards/NewCardFront.tsx
@@ -1,9 +1,9 @@
 import React, { useRef } from 'react';
 
-import styles from './Cards.module.css';
-import { handleTextareaInput } from './CardBack.tsx';
+import styles from './NewCards.module.css';
+import { handleTextareaInput } from './EditCardBack.tsx';
 
-export const CardFront = () => {
+export const NewCardFront = () => {
   const textareaRef = useRef<HTMLTextAreaElement>(null);
 
   return (
